Fix tests and allow dragging into empty list
