+++ entry/entry-complete.js
@@ -3,8 +3,8 @@ import Swap from '../plugins/Swap';
 import MultiDrag from '../plugins/MultiDrag';
 import CancelDrag from '../plugins/CancelDrag';
 
-Sortable.mount(new Swap());
+// Sortable.mount(new Swap());
 Sortable.mount(new MultiDrag());
-Sortable.mount(new CancelDrag());
+// Sortable.mount(new CancelDrag());
 
 export default Sortable;
