+++ entry/entry-core.js
@@ -3,6 +3,7 @@ import AutoScroll from '../plugins/AutoScroll';
 import OnSpill from '../plugins/OnSpill';
 import Swap from '../plugins/Swap';
 import MultiDrag from '../plugins/MultiDrag';
+import CancelDrag from '../plugins/CancelDrag';
 
 export default Sortable;
 
@@ -15,5 +16,6 @@ export {
 
 	// Extra
 	Swap,
-	MultiDrag
+	MultiDrag,
+	CancelDrag
 };
