+++ modular/sortable.esm.js
@@ -3272,8 +3272,54 @@ function removeMultiDragElements() {
   });
 }
 
+function CancelDragPlugin() {
+  class CancelDrag {
+    constructor() {
+      this.defaults = {
+        cancelDrag: true
+      };
+    }
+    drop({
+      cancel,
+      dispatchSortableEvent,
+      originalEvent,
+      dragEl,
+      cloneEl
+    }) {
+      // In case the 'ESC' key was hit,
+      // the origEvent is of type 'dragEnd'.
+      if (originalEvent?.type === 'dragend') {
+        // Call revert on spill, to revert the drag
+        // using the existing algorithm.
+        this.sortable.revertOnSpill?.onSpill(...arguments);
+
+        // Undo changes on the drag element.
+        if (dragEl) {
+          // Remove ghost & chosen class.
+          dragEl.classList.remove(this.options.ghostClass);
+          dragEl.classList.remove(this.options.chosenClass);
+          dragEl.removeAttribute('draggable');
+        }
+
+        // In case of a copy, the cloneEl
+        // has to be removed again.
+        if (cloneEl) {
+          cloneEl.remove();
+        }
+
+        // Dispatch 'end' event.
+        dispatchSortableEvent('end');
+        cancel();
+      }
+    }
+  }
+  return _extends(CancelDrag, {
+    pluginName: 'cancelDrag'
+  });
+}
+
 Sortable.mount(new AutoScrollPlugin());
 Sortable.mount(Remove, Revert);
 
 export default Sortable;
-export { MultiDragPlugin as MultiDrag, Sortable, SwapPlugin as Swap };
+export { CancelDragPlugin as CancelDrag, MultiDragPlugin as MultiDrag, Sortable, SwapPlugin as Swap };
